Fix retake count in case of deletion
